Prep folder with example
